aoclib_rs

Module iter

Functions

fwd_rev_incl_range

Forward/reverse inclusive range. If start <= end, returns a forward iterator from start to end, inclusive. Else, returns a reverse iterator from end to start, inclusive.

pairwise_iter

Iterates through every pair in the input slice. Returns references to the input.

pairwise_iter_copy

Iterates through every pair in the input slice. Returns copies of the input.

pairwise_iter_i

Iterates through every pair in the input slice. Returns indices into the input.

permutation_iter_copy

Iterates through all permutations of the input. Returns copies of the input.

selection_iter

Iterates through every possible combination of selecting elements from the input. Returns references to the input.

selector_iter

Iterates through a “selector” of len elements, generating every possible selection. The idea is that you “select” the elements in some other structure at the indices where the returned Vec is true. Put another way, the returned Vec<bool>s can be looked at as counting up in binary.