aoclib_rs

Module point

Structs

Point2d

2-dimensional point. For full functionality for floats, try the ordered_float crate

Point3d

3-dimensional point. For full functionality for floats, try the ordered_float crate

PointNd

N-dimensional point. For full functionality for floats, try the ordered_float crate

Slope

A simplified slope between two integer points.

Traits

PointDist

A point capable of computing the straight-line distance between itself and another point of the same type.

PointFromArray

A point that can be initialized with an array of the appropriate size.

PointManhattan

A point capable of computing the Manhattan distance between itself and another point of the same type.

Functions

get_angle

Returns the angle (from straight up, going clockwise) represented by this slope in radians.