aoclib_rs::point

Trait PointManhattan

pub trait PointManhattan<T> {
    // Required method
    fn manhattan(&self, other: &Self) -> T;
}

A point capable of computing the Manhattan distance between itself and another point of the same type.

Required Methods

fn manhattan(&self, other: &Self) -> T

Computes the Manhattan distance between self and other.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.